% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{root_colors}
\alias{root_colors}
\title{Colors from Leder Games' Root, specifically the art of Kyle Ferrin, extracted
from the Steam workshop mod files}
\format{
An object of class \code{character} of length 11.
}
\usage{
root_colors
}
\description{
Colors from Leder Games' Root, specifically the art of Kyle Ferrin, extracted
from the Steam workshop mod files
}
\keyword{datasets}
